requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
requests==2.32.3
ffmpeg-python==0.2.0
gTTS==2.5.1
